refactor(lsp): break out code lens logic
